[Chore] Move the config tests into the modules they test
The facade was still carrying all 32 config tests, so it stayed the
largest file in the module and no test sat next to the code it covers.
Each now lives with its subject: parsing and defaults in `schema`, file
and `opendal://` sources in `load`, `${env:}` and `HARUKI_*` handling in
`env`, every rejection case in `validate`, and pool widening in `tuning`.

`env_lock` and `restore_env` move to a `test_support` module rather than
being copied. The lock exists because these tests mutate process
environment variables; duplicating it per module would have produced two
mutexes that do not exclude each other, which fails only under load.

Same 179 tests before and after. `config.rs` is now 15 lines.
