+++ src/core/config/load.rs
@@ -248,3 +248,239 @@ fn candidate_paths() -> Vec<PathBuf> {
     candidates.push(PathBuf::from("../../haruki-asset-configs.yaml"));
     candidates
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use super::super::schema::{
+        default_asset_studio_export_types, AssetHttpVersion, BackendsConfig, ImageBackend,
+        ImagePngCompression, MediaBackend,
+    };
+    use super::super::test_support::{env_lock, restore_env};
+
+    #[test]
+    fn rejects_non_v3_config_version() {
+        let config = AppConfig {
+            config_version: 1,
+            ..AppConfig::default()
+        };
+        let err = config.validate().unwrap_err();
+        assert!(matches!(err, ConfigError::UnsupportedVersion(1)));
+    }
+
+    #[test]
+    fn parses_v3_yaml_structure() {
+        let yaml = r#"
+config_version: 3
+server:
+  host: 127.0.0.1
+  port: 18080
+  asset_http_version: http1
+  auth:
+    enabled: true
+    bearer_token: secret
+logging:
+  level: DEBUG
+execution:
+  retry:
+    attempts: 3
+    initial_backoff_ms: 250
+    max_backoff_ms: 1000
+regions:
+  jp:
+    enabled: true
+    provider:
+      kind: colorful_palette
+      asset_info_url_template: "https://example.com/{env}/{asset_version}/{asset_hash}"
+      asset_bundle_url_template: "https://example.com/assets/{bundle_path}"
+      profile: production
+      profile_hashes:
+        production: abc123
+"#;
+
+        let config: AppConfig = yaml_serde::from_str(yaml).unwrap();
+        config.validate().unwrap();
+
+        assert_eq!(config.server.port, 18080);
+        assert_eq!(config.server.asset_http_version, AssetHttpVersion::Http1);
+        assert_eq!(config.logging.level, "DEBUG");
+        assert_eq!(config.execution.retry.attempts, 3);
+        assert_eq!(config.enabled_regions(), vec!["jp".to_string()]);
+        assert_eq!(
+            config.regions["jp"].export.asset_studio_types,
+            default_asset_studio_export_types()
+        );
+    }
+
+    #[test]
+    fn parses_asset_studio_options() {
+        let yaml = r#"
+media:
+  backend: ffi
+  ffmpeg_path: ffmpeg
+image:
+  backend: rust
+  png_compression: best
+  webp_lossless: true
+  jpeg_quality: 88
+asset_studio:
+  read_batch_size: 16
+  image_format: raw_rgba
+  read_kinds:
+    Sprite: image
+    Animator: fbx
+    all: typetree_json
+"#;
+        let backends: BackendsConfig = yaml_serde::from_str(yaml).unwrap();
+        let asset_studio = &backends.asset_studio;
+        assert_eq!(backends.media.backend, MediaBackend::Ffi);
+        assert_eq!(backends.image.backend, ImageBackend::Rust);
+        assert_eq!(backends.image.png_compression, ImagePngCompression::Best);
+        assert_eq!(backends.image.jpeg_quality, 88);
+        assert_eq!(asset_studio.read_batch_size, 16);
+        assert_eq!(asset_studio.image_format.as_deref(), Some("raw_rgba"));
+        assert_eq!(
+            asset_studio.read_kinds.get("Animator").map(String::as_str),
+            Some("fbx")
+        );
+        assert_eq!(
+            asset_studio.read_kinds.get("all").map(String::as_str),
+            Some("typetree_json")
+        );
+    }
+
+    #[test]
+    fn example_config_advertises_current_haruki_3d_pipeline_selectors() {
+        let config_path =
+            Path::new(env!("CARGO_MANIFEST_DIR")).join("haruki-asset-configs.example.yaml");
+        let config = AppConfig::load_from_path(config_path).unwrap();
+        let asset_studio = &config.backends.asset_studio;
+        assert_eq!(asset_studio.read_kinds.get("Animator"), None);
+        assert_eq!(
+            asset_studio.read_kinds.get("AnimationClip"),
+            None,
+            "unconfigured types should use unity-rs TypeTree/raw fallback"
+        );
+
+        let jp = config.regions.get("jp").expect("jp region exists");
+        assert!(
+            jp.export
+                .asset_studio_types
+                .iter()
+                .any(|value| value.eq_ignore_ascii_case("all")),
+            "jp asset_studio_types should request every unity-rs readable object"
+        );
+        assert_eq!(jp.filters.start_app, vec![".*"]);
+        assert_eq!(jp.filters.on_demand, vec![".*"]);
+
+        assert!(
+            jp.export.raw_bundles.is_none(),
+            "the default example must not enable legacy raw bundle retention"
+        );
+
+        let haruki_3d = &jp.export.haruki_3d;
+        for expected in [
+            "live_pv/model/characterv2/body/",
+            "live_pv/model/characterv2/face/",
+            "live_pv/model/characterv2/head_optional/",
+            "live_pv/model/characterv2/color_variation/body/",
+            "live_pv/model/characterv2/color_variation/face/",
+            "live_pv/model/characterv2/color_variation/head_optional/",
+            "live_pv/model/character/head_optional/",
+            "live_pv/model/character/color_variation/head_optional/",
+            "character/motion/costume_setting/",
+        ] {
+            assert!(
+                haruki_3d
+                    .include
+                    .iter()
+                    .any(|value| value.contains(expected)),
+                "haruki_3d.include should retain {expected}"
+            );
+        }
+
+        assert!(
+            haruki_3d.master_dir.contains("haruki-sekai-master/master"),
+            "haruki_3d.master_dir should point at the upstream masterdata checkout"
+        );
+        assert!(
+            haruki_3d.output_dir.contains("3d-output"),
+            "haruki_3d.output_dir should point at a stable runtime root"
+        );
+        assert!(
+            haruki_3d.manifest_file.contains("3d-output"),
+            "haruki_3d.manifest_file should live beside the stable runtime root"
+        );
+        assert!(
+            haruki_3d.role_character3d_ids.contains(&5),
+            "haruki_3d.role_character3d_ids should include a v1 smoke role runtime"
+        );
+        assert_eq!(
+            haruki_3d.process_concurrency, 0,
+            "haruki_3d.process_concurrency should default to exporter auto in the example config"
+        );
+        for expected in [
+            "live_pv/model/characterv2/body/",
+            "live_pv/model/characterv2/face/",
+            "live_pv/model/characterv2/head_optional/",
+            "live_pv/model/characterv2/color_variation/body/",
+            "live_pv/model/characterv2/color_variation/face/",
+            "live_pv/model/characterv2/color_variation/head_optional/",
+            "live_pv/model/character/head_optional/",
+            "live_pv/model/character/color_variation/head_optional/",
+            "character/motion/costume_setting/",
+        ] {
+            assert!(
+                haruki_3d
+                    .include
+                    .iter()
+                    .any(|value| value.contains(expected)),
+                "haruki_3d.include should stage {expected}"
+            );
+        }
+    }
+
+    #[tokio::test]
+    #[allow(clippy::await_holding_lock)]
+    async fn load_default_reads_config_from_opendal_fs_uri() {
+        let _env_lock = env_lock();
+        let dir = tempfile::tempdir().unwrap();
+        std::fs::write(
+            dir.path().join("haruki-asset-configs.yaml"),
+            r#"
+config_version: 3
+server:
+  port: 19090
+regions:
+  cn:
+    enabled: true
+    provider:
+      kind: nuverse
+      asset_version_url: "https://example.com/version"
+      app_version: "5.2.0"
+      asset_info_url_template: "https://example.com/info/{asset_version}"
+      asset_bundle_url_template: "https://example.com/{bundle_path}"
+"#,
+        )
+        .unwrap();
+
+        let old_config_uri = std::env::var("HARUKI_CONFIG_URI").ok();
+        let old_scheme = std::env::var("HARUKI_CONFIG_OPENDAL_SCHEME").ok();
+        let old_root = std::env::var("HARUKI_CONFIG_OPENDAL_ROOT").ok();
+        std::env::set_var(
+            "HARUKI_CONFIG_URI",
+            "opendal://config/haruki-asset-configs.yaml",
+        );
+        std::env::set_var("HARUKI_CONFIG_OPENDAL_SCHEME", "fs");
+        std::env::set_var("HARUKI_CONFIG_OPENDAL_ROOT", dir.path());
+
+        let config = AppConfig::load_default().await.unwrap();
+        assert_eq!(config.server.port, 19090);
+        assert_eq!(config.enabled_regions(), vec!["cn".to_string()]);
+
+        restore_env("HARUKI_CONFIG_URI", old_config_uri);
+        restore_env("HARUKI_CONFIG_OPENDAL_SCHEME", old_scheme);
+        restore_env("HARUKI_CONFIG_OPENDAL_ROOT", old_root);
+    }
+}
